hello-ai/requirements.txt
# Core AI packages
transformers==4.56.0
torch>=2.0.0
huggingface-hub>=0.34.0
tokenizers>=0.22.0
safetensors>=0.6.0

# API clients
anthropic==0.66.0
pygeai-chat>=2025.05.0

# Utilities
python-dotenv>=1.0.0
numpy>=1.17.0
regex>=2020.1.8
requests>=2.25.0
tqdm>=4.27.0
PyYAML>=5.1.0

# HTTP and async support
httpx>=0.25.0
anyio>=3.5.0

# Core dependencies (will be installed automatically but listed for clarity)
packaging>=20.0
filelock>=3.0.0
